Add support for enabled topic and improve topic configuration
Instead of manually changing the topic strings, they are now generated
at compile-time using a C89 feature together with a define that contains
the base topic.

Also added support for enabling/disabling the device remotely by
subscribing to a `/enabled` topic and improved handling of the topics
in the MQTT callback.