Add MPU6050 support and fix null termination bug
As the goal is to stram MPU movement events we need to interface with
the sensor. This commit adds initial support for taking accelerometer,
gyro and temperature measurements from the MPU6050 sensor and publish
them as JSON formatted strings on MQTT.

A bug was found that caused the topic callback to fail for the enabled
topic. The callback is called with an argument specifying the length of
the payload buffer. Before, we were blindly passing the raw payload to
the String constructor, causing it to also account for old received \
messages when creating the string. The fix insserts a null-termination
character at the position specified by the payload lenght. This way the
String constructor correctly hits the end of the payload. Note that This
solution is only valid as long as `len` is less than the length of the
receiving buffer. Otherwise, we will write outside the array into other
memory. As the payloads are so short (for now at least) this is considered
a valid risk to take for a simple solution. A more robust solution would
be to copy the payload contents into a separate buffer that is at least
`len + 1` long.