Add class for handling button presses
Adding a button and button LED to the project in order to enable and
disable the measuring directly on the device. So far only a class that
handles the button short and long presses has been added, but more to come!

Also changed the previous LED `const` variable to a `define`.